+++ internal/controller/operator/factory/vtcluster/cluster_test.go
@@ -508,4 +508,82 @@ func TestCreateOrUpdate(t *testing.T) {
 			assert.True(t, k8serrors.IsNotFound(err))
 		},
 	})
+
+	// vtcluster with load-balancing
+	f(opts{
+		cr: &vmv1.VTCluster{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "test",
+				Namespace: "default",
+			},
+			Spec: vmv1.VTClusterSpec{
+				RequestsLoadBalancer: vmv1beta1.VMAuthLoadBalancer{
+					Enabled: true,
+					Spec: vmv1beta1.VMAuthLoadBalancerSpec{
+						CommonAppsParams: vmv1beta1.CommonAppsParams{
+							ReplicaCount: ptr.To(int32(1)),
+						},
+					},
+				},
+				Insert: &vmv1.VTInsert{
+					CommonAppsParams: vmv1beta1.CommonAppsParams{ReplicaCount: ptr.To(int32(0))},
+				},
+				Select: &vmv1.VTSelect{
+					CommonAppsParams: vmv1beta1.CommonAppsParams{ReplicaCount: ptr.To(int32(0))},
+				},
+				Storage: &vmv1.VTStorage{
+					CommonAppsParams: vmv1beta1.CommonAppsParams{ReplicaCount: ptr.To(int32(0))},
+				},
+			},
+		},
+		validate: func(ctx context.Context, rclient client.Client, cr *vmv1.VTCluster) {
+			var dep appsv1.Deployment
+			assert.NoError(t, rclient.Get(ctx, types.NamespacedName{
+				Namespace: cr.Namespace,
+				Name:      cr.PrefixedName(vmv1beta1.ClusterComponentBalancer),
+			}, &dep))
+		},
+	})
+
+	// vtcluster with load-balancing and HPA
+	f(opts{
+		cr: &vmv1.VTCluster{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "test",
+				Namespace: "default",
+			},
+			Spec: vmv1.VTClusterSpec{
+				RequestsLoadBalancer: vmv1beta1.VMAuthLoadBalancer{
+					Enabled: true,
+					Spec: vmv1beta1.VMAuthLoadBalancerSpec{
+						CommonAppsParams: vmv1beta1.CommonAppsParams{
+							ReplicaCount: ptr.To(int32(1)),
+						},
+						HPA: &vmv1beta1.EmbeddedHPA{
+							MinReplicas: ptr.To(int32(1)),
+							MaxReplicas: 3,
+						},
+					},
+				},
+				Insert: &vmv1.VTInsert{
+					CommonAppsParams: vmv1beta1.CommonAppsParams{ReplicaCount: ptr.To(int32(0))},
+				},
+				Select: &vmv1.VTSelect{
+					CommonAppsParams: vmv1beta1.CommonAppsParams{ReplicaCount: ptr.To(int32(0))},
+				},
+				Storage: &vmv1.VTStorage{
+					CommonAppsParams: vmv1beta1.CommonAppsParams{ReplicaCount: ptr.To(int32(0))},
+				},
+			},
+		},
+		validate: func(ctx context.Context, rclient client.Client, cr *vmv1.VTCluster) {
+			var hpa autoscalingv2.HorizontalPodAutoscaler
+			assert.NoError(t, rclient.Get(ctx, types.NamespacedName{
+				Namespace: cr.Namespace,
+				Name:      cr.PrefixedName(vmv1beta1.ClusterComponentBalancer),
+			}, &hpa))
+			assert.Equal(t, int32(3), hpa.Spec.MaxReplicas)
+			assert.Equal(t, "Deployment", hpa.Spec.ScaleTargetRef.Kind)
+		},
+	})
 }